requirements.txt
fastapi
uvicorn
pymongo
pinecone-client
sentence-transformers
python-dotenv
certifi
